+++ internal/sonos/didl_test.go
@@ -36,3 +36,29 @@ func TestParseDIDLItems_Minimal(t *testing.T) {
 		t.Fatalf("unexpected id: %q", items[0].ID)
 	}
 }
+
+func TestParseDIDLItems_ResMD(t *testing.T) {
+	t.Parallel()
+
+	didl := `<DIDL-Lite xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:r="urn:schemas-rinconnetworks-com:metadata-1-0/" xmlns="urn:schemas-upnp-org:metadata-1-0/DIDL-Lite/">
+<item id="FV:2/1" parentID="FV:2" restricted="true">
+  <dc:title>Favorite</dc:title>
+  <res>x-rincon-mp3radio:http://example.com/stream</res>
+  <r:resMD>&lt;DIDL-Lite&gt;&lt;item&gt;&lt;dc:title&gt;X&lt;/dc:title&gt;&lt;/item&gt;&lt;/DIDL-Lite&gt;</r:resMD>
+</item>
+</DIDL-Lite>`
+
+	items, err := ParseDIDLItems(didl)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	if len(items) != 1 {
+		t.Fatalf("expected 1 item, got %d", len(items))
+	}
+	if items[0].ResMD == "" {
+		t.Fatalf("expected resMD to be parsed")
+	}
+	if items[0].ResMD[0] != '<' {
+		t.Fatalf("expected unescaped resMD, got: %q", items[0].ResMD)
+	}
+}